# Go dependency download and analysis performance fix (Cherry-pick of #22726)