--- a/src/python/pants/backend/go/util_rules/third_party_pkg.py
+++ b/src/python/pants/backend/go/util_rules/third_party_pkg.py
@@ -37,7 +37,7 @@ from pants.engine.fs import (
     Snapshot,
 )
 from pants.engine.process import FallibleProcessResult, Process, ProcessResult
-from pants.engine.rules import Get, MultiGet, collect_rules, concurrently, rule
+from pants.engine.rules import Get, MultiGet, collect_rules, concurrently, implicitly, rule
 from pants.util.dirutil import group_by_dir
 from pants.util.frozendict import FrozenDict
 from pants.util.logging import LogLevel
@@ -599,6 +599,7 @@ async def download_and_analyze_third_party_packages(
                 minimum_go_version=mod.minimum_go_version,
                 build_opts=request.build_opts,
             ),
+            **implicitly(),
         )
         for mod in module_analysis.modules
     )